refactor and clean code
- remove unused PlayerGameProps type inside PlayInfo
- merge board's _default and game's _data into lib/default.ts
- move board's utils to lib/utils/board
