+++ src/lib/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./generate";
 export * from "./api";
 export * from "./general";
 export * from "./chess";
+export * from "./board";
